+++ ChangeLog.md
@@ -1,3 +1,11 @@
-# Changelog for transdoc
+# Changelog for FromHTML
 
-## Unreleased changes
+## [1.0.1] - 2018-08-19
+
+### Changed
+- Getting rid of ugly unsafe performing IO
+- Using process-extra with ByteString to avoid encoding problems
+
+## [1.0.0] - 2018-08-09
+
+Initial release